backend_service/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.22
psycopg2-binary==2.9.6
pydantic==1.10.12
requests==2.31.0
python-dotenv==1.0.0
httpx==0.24.1
asyncpg==0.27.0
SQLAlchemy[asyncio]==2.0.22
python-dateutil==2.8.2
redis==4.7.0
boto3==1.29.0
botocore==1.36.0
mangum==0.14.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
